LayoutTests/fast/mediastream/constructors.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests the JSEP PeerConnection related constructors.");

shouldBeTrue("typeof webkitRTCPeerConnection === 'function'");
shouldBeTrue("typeof RTCSessionDescription === 'function'");
shouldBeTrue("typeof RTCIceCandidate === 'function'");

shouldThrow("webkitRTCPeerConnection()");
shouldThrow("RTCSessionDescription()");
shouldThrow("RTCIceCandidate()");

shouldNotThrow("new webkitRTCPeerConnection({iceServers:[{url:'stun://foobar.com:12345'}]}, null);");
shouldNotThrow("new RTCSessionDescription({type:'offer',sdp:'foobar'});");
shouldNotThrow("new RTCIceCandidate({candidate:'foo'});");

window.jsTestIsAsync = false;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
